Fixed an issue where a user is trying to Zap Out of a vault. Same occurs with Zap In.
If the user hasn't approved yet, the UI queues both the approval transaction and the deposit transaction immediately afterwards.
But the zapper API doesn't pick up that allowance has been increased yet, so it returns an error.

I wait for the approve call to be mined, + 1 confirmation before queuing the withdraw transaction.

Signed-off-by: antonnell <antonnell.crpto@gmail.com>