:construction_worker: Setup GitHub Action CI
Enables local build and linting as part of Continuous Integration.
The linter currently ignores a lot of rules. Theses rules will be
enabled as we later fix them.
Also fixes leftover linting issues in `vaultAssetRow.js`.
Note that `NEXT_PUBLIC_PROVIDER` is expected to be set for the build,
but the actual value doesn't currently matter.